PRDBench/5/src/requirements.txt
opencv-python>=4.5.0
Pillow>=9.0.0
scikit-learn>=1.0.0
numpy>=1.20.0
requests>=2.25.0
